Fix some clang-tidy bugprone-narrowing-conversions warnings

+++ src/video/GLImage.hh
@@ -19,9 +19,9 @@ public:
 	GLImage(OutputSurface& output, SDLSurfacePtr image);
 	GLImage(OutputSurface& output, const std::string& filename, float scaleFactor);
 	GLImage(OutputSurface& output, const std::string& filename, gl::ivec2 size);
-	GLImage(OutputSurface& output, gl::ivec2 size, unsigned rgba);
-	GLImage(OutputSurface& output, gl::ivec2 size, std::span<const unsigned, 4> rgba,
-	        int borderSize, unsigned borderRGBA);
+	GLImage(OutputSurface& output, gl::ivec2 size, uint32_t rgba);
+	GLImage(OutputSurface& output, gl::ivec2 size, std::span<const uint32_t, 4> rgba,
+	        int borderSize, uint32_t borderRGBA);
 
 	void draw(OutputSurface& output, gl::ivec2 pos,
 	          uint8_t r, uint8_t g, uint8_t b, uint8_t alpha) override;
@@ -34,8 +34,8 @@ private:
 	gl::BufferObject elementsBuffer;
 	gl::Texture texture; // must come after size
 	int borderSize;
-	std::array<int, 4> bgA;
-	int borderA;
+	std::array<uint16_t, 4> bgA; // 0..256
+	uint16_t borderA; // 0..256
 	std::array<uint8_t, 4> bgR, bgG, bgB;
 	uint8_t borderR, borderG, borderB;
 };
